# openapi-schema.yml
openapi: 3.0.3
info:
  title: NoCode Backend API
  version: 1.0.0
  description: API pour la plateforme NoCode - Création d'applications sans code
paths:
...
  /api/v1/automation/integrations/:
    get:
      operationId: automation_integrations_list
      description: |-
        API pour gérer les intégrations.

        Endpoints:
        - GET /api/automation/integrations/ : Liste des intégrations
        - GET /api/automation/integrations/available/ : Intégrations disponibles
        - POST /api/automation/integrations/configure/ : Configurer une intégration
        - POST /api/automation/integrations/{id}/test/ : Tester une intégration
        - POST /api/automation/integrations/{id}/execute/ : Exécuter une action
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIntegrationList'
          description: ''
    post:
      operationId: automation_integrations_create
      description: |-
        API pour gérer les intégrations.

        Endpoints:
        - GET /api/automation/integrations/ : Liste des intégrations
        - GET /api/automation/integrations/available/ : Intégrations disponibles
        - POST /api/automation/integrations/configure/ : Configurer une intégration
        - POST /api/automation/integrations/{id}/test/ : Tester une intégration
        - POST /api/automation/integrations/{id}/execute/ : Exécuter une action
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IntegrationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IntegrationRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Integration'
          description: ''
...
  /api/v1/automation/workflows/:
    get:
      operationId: automation_workflows_list
      description: |-
        API pour gérer les workflows.

        Endpoints:
        - GET /api/automation/workflows/ : Liste des workflows
        - POST /api/automation/workflows/ : Créer un workflow
        - GET /api/automation/workflows/{id}/ : Détails d'un workflow
        - PUT /api/automation/workflows/{id}/ : Mettre à jour un workflow
        - DELETE /api/automation/workflows/{id}/ : Supprimer un workflow
        - POST /api/automation/workflows/{id}/execute/ : Exécuter un workflow
        - GET /api/automation/workflows/{id}/executions/ : Historique d'exécutions
        - POST /api/automation/workflows/{id}/validate/ : Valider un workflow
        - POST /api/automation/workflows/{id}/duplicate/ : Dupliquer un workflow
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWorkflowListList'
          description: ''
    post:
      operationId: automation_workflows_create
      description: |-
        API pour gérer les workflows.

        Endpoints:
        - GET /api/automation/workflows/ : Liste des workflows
        - POST /api/automation/workflows/ : Créer un workflow
        - GET /api/automation/workflows/{id}/ : Détails d'un workflow
        - PUT /api/automation/workflows/{id}/ : Mettre à jour un workflow
        - DELETE /api/automation/workflows/{id}/ : Supprimer un workflow
        - POST /api/automation/workflows/{id}/execute/ : Exécuter un workflow
        - GET /api/automation/workflows/{id}/executions/ : Historique d'exécutions
        - POST /api/automation/workflows/{id}/validate/ : Valider un workflow
        - POST /api/automation/workflows/{id}/duplicate/ : Dupliquer un workflow
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WorkflowCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WorkflowCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreate'
          description: ''
...
  /api/v1/automation/workflows/{id}/edges/:
    get:
      operationId: automation_workflows_edges_list
      description: ViewSet pour les opérations CRUD sur les arêtes.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEdgeList'
          description: ''
    post:
      operationId: automation_workflows_edges_create
      description: ViewSet pour les opérations CRUD sur les arêtes.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EdgeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EdgeRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edge'
          description: ''
...
  /api/v1/automation/workflows/{id}/nodes/:
    get:
      operationId: automation_workflows_nodes_list
      description: ViewSet pour les opérations CRUD sur les nœuds.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNodeList'
          description: ''
    post:
      operationId: automation_workflows_nodes_create
      description: ViewSet pour les opérations CRUD sur les nœuds.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NodeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NodeRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
          description: ''
...
components:
  schemas:
...
    EdgeRequest:
      type: object
      description: Serializer pour les arêtes du graphe de workflows.
      properties:
        workflow:
          type: string
          format: uuid
        source_node:
          type: string
          format: uuid
        target_node:
          type: string
          format: uuid
        source_port:
          type: string
          minLength: 1
          default: output
          maxLength: 50
        target_port:
          type: string
          minLength: 1
          default: input
          maxLength: 50
        label:
          type: string
          maxLength: 255
      required:
      - source_node
      - target_node
      - workflow
...
    Integration:
      type: object
      description: Serializer pour les intégrations.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          title: Nom de l'intégration
          maxLength: 255
        integration_type:
          enum:
          - email
          - stripe
          - webhook
          - database
          - storage
          - notification
          - api
          - slack
          - zapier
          - custom
          type: string
          description: |-
            * `email` - Email (SMTP/SendGrid)
            * `stripe` - Stripe
            * `webhook` - Webhook HTTP
            * `database` - Base de données
            * `storage` - Stockage (S3, etc.)
            * `notification` - Notifications (SMS, Push)
            * `api` - API REST personnalisée
            * `slack` - Slack
            * `zapier` - Zapier
            * `custom` - Personnalisée
          title: Type d'intégration
        organization:
          type: integer
          title: Organisation
        organization_name:
          type: string
          readOnly: true
        config:
          type: object
          additionalProperties: {}
          title: Configuration
          description: Configuration générale de l'intégration
        status:
          enum:
          - active
          - inactive
          - error
          - testing
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `error` - Erreur
            * `testing` - En test
          title: Statut
        total_calls:
          type: integer
          readOnly: true
          title: Appels totaux
        successful_calls:
          type: integer
          readOnly: true
          title: Appels réussis
        failed_calls:
          type: integer
          readOnly: true
          title: Appels échoués
        success_rate:
          type: string
          readOnly: true
        last_used_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: Dernière utilisation
        rate_limit:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          title: Limite de requêtes
          description: Nombre maximum de requêtes par heure
        metadata:
          type: object
          additionalProperties: {}
          title: Métadonnées
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: Créé le
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: Modifié le
      required:
      - integration_type
      - name
      - organization
    IntegrationRequest:
      type: object
      description: Serializer pour les intégrations.
      properties:
        name:
          type: string
          minLength: 1
          title: Nom de l'intégration
          maxLength: 255
        integration_type:
          enum:
          - email
          - stripe
          - webhook
          - database
          - storage
          - notification
          - api
          - slack
          - zapier
          - custom
          type: string
          description: |-
            * `email` - Email (SMTP/SendGrid)
            * `stripe` - Stripe
            * `webhook` - Webhook HTTP
            * `database` - Base de données
            * `storage` - Stockage (S3, etc.)
            * `notification` - Notifications (SMS, Push)
            * `api` - API REST personnalisée
            * `slack` - Slack
            * `zapier` - Zapier
            * `custom` - Personnalisée
          title: Type d'intégration
        organization:
          type: integer
          title: Organisation
        config:
          type: object
          additionalProperties: {}
          title: Configuration
          description: Configuration générale de l'intégration
        status:
          enum:
          - active
          - inactive
          - error
          - testing
          type: string
          description: |-
            * `active` - Active
            * `inactive` - Inactive
            * `error` - Erreur
            * `testing` - En test
          title: Statut
        rate_limit:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          title: Limite de requêtes
          description: Nombre maximum de requêtes par heure
        metadata:
          type: object
          additionalProperties: {}
          title: Métadonnées
      required:
      - integration_type
      - name
      - organization
...
    PaginatedEdgeList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Edge'
...
    PaginatedIntegrationList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Integration'
...
    PaginatedWorkflowListList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowList'
...
    WorkflowCreate:
      type: object
      description: Serializer pour créer un workflow.
      properties:
        name:
          type: string
          title: Nom du workflow
          maxLength: 255
        description:
          type: string
        organization:
          type: integer
          title: Organisation
        project:
          type: integer
          nullable: true
          title: Projet
        status:
          enum:
          - DRAFT
          - ACTIVE
          - PAUSED
          - ARCHIVED
          type: string
          description: |-
            * `DRAFT` - Brouillon
            * `ACTIVE` - Actif
            * `PAUSED` - En pause
            * `ARCHIVED` - Archivé
          title: Statut
        config:
          type: object
          additionalProperties: {}
          title: Configuration
          description: Configuration générale du workflow
        variables:
          type: object
          additionalProperties: {}
          description: Variables globales du workflow
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/Trigger'
      required:
      - name
      - organization
    WorkflowCreateRequest:
      type: object
      description: Serializer pour créer un workflow.
      properties:
        name:
          type: string
          minLength: 1
          title: Nom du workflow
          maxLength: 255
        description:
          type: string
        organization:
          type: integer
          title: Organisation
        project:
          type: integer
          nullable: true
          title: Projet
        status:
          enum:
          - DRAFT
          - ACTIVE
          - PAUSED
          - ARCHIVED
          type: string
          description: |-
            * `DRAFT` - Brouillon
            * `ACTIVE` - Actif
            * `PAUSED` - En pause
            * `ARCHIVED` - Archivé
          title: Statut
        config:
          type: object
          additionalProperties: {}
          title: Configuration
          description: Configuration générale du workflow
        variables:
          type: object
          additionalProperties: {}
          description: Variables globales du workflow
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStepRequest'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/TriggerRequest'
      required:
      - name
      - organization
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.nocode.com
  description: Production server